requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-crontab==0.7.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
flake8==6.0.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
jedi==0.18.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
openapi-codec==1.3.2
packaging==23.0
parso==0.8.3
Pillow==9.4.0
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
pytz==2022.7
PyYAML==6.0
reportlab==3.6.12
requests==2.28.2
ruamel.yaml==0.17.21
simplejson==3.18.3
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
urwid==2.1.2
urwid-readline==0.13
drf-spectacular==0.25.1
pillow==9.4.0